requirements.txt
requests
aiohttp
aiopg==0.16.0
psycopg2
peewee_async==0.5.11
pillow
feedparser
lxml
cssselect
romajitable
transliterate
captionbot
googletrans
numpy
asyncio_redis
mc.py
yandex-translater
py-cpuinfo
apscheduler
sentry-sdk==0.14.1

git+https://github.com/smiley/steamapi
